Add files via upload
My own personal contribution to our Statistical Machine Learning mini-project. The project description was to try to train chosen machine learning methods on a given dataset. The task was to try and predict which attributes were of importance when classifying whether a person liked a song or not. Using this trained model, a prediction of a new dataset of songs was made.